requirements.txt
py-clob-client>=0.19.0
web3==6.14.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
pandas>=2.0.0
colorama>=0.4.6
